refactors sanitizable behaviour
